fix CROS
