+++ main.go
@@ -1,22 +1,5 @@
 package main
 
-import (
-	"os"
-
-	"github.com/gin-gonic/gin"
-	_ "github.com/go-sql-driver/mysql"
-)
-
-func setPortandGinMode() string {
-	if value, ok := os.LookupEnv("PORT"); ok {
-		gin.SetMode(gin.ReleaseMode)
-		return value
-	}
-	return "8080"
-}
-
 func main() {
-	port := setPortandGinMode()
 	initRoutes()
-	router.Run(":" + port)
 }
